Moving pagination into RecordingsList
