+++ app/javascript/components/recordings/Recordings.jsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 import {
-  Card, Button, Stack, Container,
+  Card, Button, Stack,
 } from 'react-bootstrap';
-import { Pagination } from 'semantic-ui-react';
 import useRecordings from '../../hooks/queries/recordings/useRecordings';
 import SearchBarQuery from '../shared/SearchBarQuery';
 import RecordingsList from './RecordingsList';
@@ -21,12 +20,6 @@ export default function Recordings() {
   const recordingsData = recordings.data.data;
   const recordingsMeta = recordings.data.meta;
 
-  const handlePage = (e, { activePage }) => {
-    const gotopage = { activePage };
-    const pagenum = gotopage.activePage;
-    setPage(pagenum);
-  };
-
   return (
     <div className="wide-background full-height-rooms">
       <Stack direction="horizontal" className="w-100 mt-5">
@@ -39,19 +32,10 @@ export default function Recordings() {
           recordings={recordingsData}
           isLoading={recordings.isLoading}
           RecordingRow={RoomsRecordingRow}
+          setPage={setPage}
+          recordingsMeta={recordingsMeta}
         />
       </Card>
-      <Container className="text-center">
-        <Pagination
-          defaultActivePage={recordingsMeta.page}
-          totalPages={recordingsMeta.pages}
-          onPageChange={handlePage}
-          firstItem={null}
-          lastItem={null}
-          pointing
-          secondary
-        />
-      </Container>
     </div>
   );
 }
